ai_service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pymongo
numpy
python-dotenv
